shared/playdevs/STORIES/CEB-002/OUTPUTS/cebraspe-parser/src/parser/ocr_engine.py
"""
OCR Engine

Este módulo é responsável pelo reconhecimento óptico de caracteres
em PDFs baseados em imagem.
Responsabilidades:
- Conversão de PDF para imagens
- Pré-processamento de imagens
- OCR com Tesseract
- Pós-processamento de texto
"""

import pytesseract
import fitz  # PyMuPDF
from PIL import Image
import cv2
import numpy as np
from pathlib import Path
from typing import List, Optional
import time

# TODO: Implementar imports necessários
# from src.config.settings import ProcessingConfig
# from src.models.document import ExtractedText
# from src.utils.logger import get_logger

class OCREngine:
    """
    Motor de OCR
    
    Responsável por realizar OCR em PDFs baseados em imagem,
    incluindo pré e pós-processamento.
    """
    
    def __init__(self, config):
        """
        Inicializa o motor de OCR
        
        Args:
            config: Configurações do processamento
        """
        self.config = config
        # self.logger = get_logger(__name__)
        
        # Configurar Tesseract
        tesseract_config = config.get_tesseract_config()
        pytesseract.pytesseract.tesseract_cmd = tesseract_config["cmd"]
    
    def extract_text_ocr(self, file_path: str) -> ExtractedText:
        """
        Extrai texto usando OCR
        
        Args:
            file_path: Caminho do arquivo PDF
            
        Returns:
            ExtractedText: Texto extraído com metadados
            
        Raises:
            Exception: Se OCR falhar
        """
        try:
            start_time = time.time()
            print(f"👁️  Executando OCR: {file_path}")
            
            # TODO: Implementar OCR real
            # 1. Converter PDF para imagens
            # 2. Pré-processar imagens
            # 3. Executar OCR
            # 4. Pós-processar texto
            # 5. Calcular confiança
            
            # Placeholder - implementar lógica real
            extracted_text = "Texto extraído via OCR (placeholder)"
            page_count = 10  # Placeholder
            confidence_score = 0.85  # Placeholder
            
            processing_time = time.time() - start_time
            
            result = ExtractedText.create(
                original_file=file_path,
                extracted_text=extracted_text,
                page_count=page_count,
                extraction_method="ocr",
                confidence_score=confidence_score,
                processing_time=processing_time
            )
            
            print(f"✅ OCR concluído: {result.word_count} palavras")
            return result
            
        except Exception as e:
            # self.logger.error(f"Erro no OCR de {file_path}: {e}")
            raise
    
    def _pdf_to_images(self, file_path: str) -> List[Image.Image]:
        """
        Converte PDF para lista de imagens
        
        Args:
            file_path: Caminho do arquivo PDF
            
        Returns:
            List[Image.Image]: Lista de imagens
        """
        # TODO: Implementar conversão real
        # 1. Abrir PDF com PyMuPDF
        # 2. Converter cada página para imagem
        # 3. Aplicar DPI configurado
        # 4. Retornar lista de imagens
        
        return []  # Placeholder
    
    def _preprocess_image(self, image: Image.Image) -> Image.Image:
        """
        Pré-processa imagem para OCR
        
        Args:
            image: Imagem original
            
        Returns:
            Image.Image: Imagem processada
        """
        # TODO: Implementar pré-processamento real
        # 1. Converter para escala de cinza
        # 2. Aplicar filtros
        # 3. Melhorar contraste
        # 4. Reduzir ruído
        
        return image  # Placeholder
    
    def _perform_ocr(self, image: Image.Image) -> tuple:
        """
        Executa OCR em uma imagem
        
        Args:
            image: Imagem para processar
            
        Returns:
            tuple: (texto, confiança)
        """
        # TODO: Implementar OCR real
        # 1. Configurar Tesseract
        # 2. Executar OCR
        # 3. Extrair texto e confiança
        # 4. Retornar resultados
        
        return "Texto OCR (placeholder)", 0.85  # Placeholder
    
    def _postprocess_text(self, text: str) -> str:
        """
        Pós-processa texto OCR
        
        Args:
            text: Texto bruto do OCR
            
        Returns:
            str: Texto processado
        """
        # TODO: Implementar pós-processamento real
        # 1. Corrigir erros comuns
        # 2. Normalizar espaçamento
        # 3. Remover artefatos
        # 4. Aplicar regras específicas
        
        return text  # Placeholder









