shared/playdevs/STORIES/CEB-002/OUTPUTS/cebraspe-parser/requirements.txt
# Core dependencies
PyPDF2>=3.0.0
pdfplumber>=0.9.0
pymupdf>=1.23.0

# OCR dependencies
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0

# Text processing
nltk>=3.8.0
spacy>=3.7.0
regex>=2023.0.0

# Data handling
pydantic>=2.0.0
python-dateutil>=2.8.0
chardet>=5.2.0

# Async support
asyncio>=3.4.3
aiofiles>=23.0.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
bandit>=1.7.0

# Machine Learning (for classification)
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0


